Fixes failing test on macOS, preventing install (#5)
* Fixes failing test on macOS, preventing install

Test `11keywords.t` fails on recent versions of macOS due to Apple
relocating files. This commit test's for `darwin` and adds the necessary
prefix.

See also @hakonhagland comment in #2.

Closes #2

* Removes hard coded perl shebang

Wouldn't want to use the wrong perl if we're running tests outside of
the test rig.

* Reworks fix for failing test caused by Apple's shenanigans

Looks for and opens `keywords.h` in a subroutine, passing out a
file handle. First we build a little list of locations, starting with
our default and expected location. Next we special case a location if
we're on macOS and the first location doesn't exist. At this point it
would be trivial to add more locations if we wanted. Frankly the check
for the existence of our preferred location is extraneous, as the best
way to see if a file exists (and is readable) is to try and open it.

* Uses @rurban's preferred 'on-darwin-and-file-missing' test